+++ src/main/java/com/supermarket/service/CheckoutService.java
@@ -3,10 +3,11 @@ package com.supermarket.service;
 import com.supermarket.enums.PromotionStrategy;
 import com.supermarket.vo.PurchaseItem;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 //结账服务接口
 public interface CheckoutService {
 
-   double checkout(List<PurchaseItem> items, PromotionStrategy promotion);
+   BigDecimal checkout(List<PurchaseItem> items, PromotionStrategy promotion);
 }
